+++ packages/components/src/components/scroller/scroller-styles.spec.ts
@@ -2,16 +2,16 @@ import { getComponentCss } from './scroller-styles';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
-    ['background-base', false, false, 'center', 'light'],
-    ['background-surface', false, false, 'center', 'dark'],
-    ['background-surface', false, false, 'center', 'light'],
-    ['background-base', false, false, 'center', 'dark'],
-    ['background-base', false, false, 'top', 'light'],
-    ['background-base', true, false, undefined, 'light'],
-    ['background-base', false, true, undefined, 'dark'],
-    ['background-base', true, true, undefined, 'light'],
+    ['background-base', false, false, 'center', false, 'light'],
+    ['background-surface', false, false, 'center', false, 'dark'],
+    ['background-surface', false, false, 'center', false, 'light'],
+    ['background-base', false, false, 'center', false, 'dark'],
+    ['background-base', false, false, 'top', true, 'light'],
+    ['background-base', true, false, undefined, true, 'light'],
+    ['background-base', false, true, undefined, true, 'dark'],
+    ['background-base', true, true, undefined, true, 'light'],
   ])(
-    'should return correct css for gradientColor: %s, isNextHidden: %s, isPrevHidden: %s, scrollIndicatorPosition: %s and theme: %s',
+    'should return correct css for gradientColor: %s, isNextHidden: %s, isPrevHidden: %s, scrollIndicatorPosition: %s, hasScrollbar: %s and theme: %s',
     (...args) => {
       expect(getComponentCss(...args)).toMatchSnapshot();
     }