+++ packages/components/src/components/table/table-head/table-head-styles.spec.ts
@@ -1,6 +1,6 @@
 import { getComponentCss } from './table-head-styles';
 
-xdescribe('getComponentCss()', () => {
+describe('getComponentCss()', () => {
   it('should return correct css', () => {
     expect(getComponentCss()).toMatchSnapshot();
   });